Update package version and enhance code quality checks
This commit increases the version of "@elsikora/dapigate-types" in package.json and expands the scope of Prettier to include all .ts files. It also includes enhancements by adding security scans, mirroring to CodeCommit, and deleting an abstract module in the changeset. Qodana's GitHub Workflow file was renamed and modified to run only on push events instead of multiple triggers.